requirements.txt
fastapi
uvicorn
python-multipart
Jinja2
uvicorn[standard]
gunicorn
pyyaml
toml